backend/requirements.txt
Flask
Flask-CORS
python-dotenv
PyJWT
supabase
praw
feedparser